backend/requirements.txt
django
djangorestframework
requests
djangorestframework-simplejwt
PyJWT
django-cors-headers
psycopg2
sqlparse
python-dotenv
Pillow
nepse-api